feat: cache mint recommendations per mode, add reviews modal to MintDiscovery, extract MintReviewsModal component
